third_party/blink/web_tests/external/wpt/css/css-ui/appearance-initial-value-001-expected.txt

This is a testharness.js-based test.
FAIL support for appearance assert_equals: expected "button" but got ""
FAIL initial value for appearance assert_equals: expected "none" but got ""
PASS support for -webkit-appearance
PASS initial value for -webkit-appearance
Harness: the test ran to completion.
